tg_bot/requirements.txt
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-telegram-bot==20.3
httpx==0.24.1